+++ src/Games/index.js
@@ -1,0 +1,10 @@
+import { rockPaperScissor } from "./rockPaperScissor.js";
+
+/**
+ * Here are some complete games logics implemented
+ * @namespace Games
+ */
+
+export {
+	rockPaperScissor
+};