+++ src/index.js
@@ -1,4 +1,5 @@
 import * as DOM from "./DOM/index.js";
+import * as Games from "./Games/index.js";
 
 import { allCharactersSame } from "./allCharactersSame.js";
 import { allEqual } from "./allEqual.js";
@@ -54,7 +55,6 @@ import { randomRGBAColor } from "./randomRGBAColor.js";
 import { randomRGBColor } from "./randomRGBColor.js";
 import { randomString } from "./randomString.js";
 import { range } from "./range.js";
-import { rockPaperScissor } from "./rockPaperScissor.js";
 import { stringToScreamingSnakeCase, screamingSnakeCaseToNormal } from "./screamingSnakeCase.js";
 import { stringToSnakeCase, snakeCaseToNormal } from "./snakeCase.js";
 import { sortAscending } from "./sortAscending.js";
@@ -70,6 +70,7 @@ import { xor } from "./xor.js";
 
 export {
 	DOM,
+	Games,
 	allCharactersSame,
 	allEqual,
 	and,
@@ -124,7 +125,6 @@ export {
 	randomRGBColor,
 	randomString,
 	range,
-	rockPaperScissor,
 	stringToScreamingSnakeCase, screamingSnakeCaseToNormal,
 	stringToSnakeCase, snakeCaseToNormal,
 	sortAscending,